refactor: split core result aggregation
